Remove ensstat task from reforecast (#3143)
This PR removes the ensstat task from the reforecast. 

This task is removed due to an occasional segmentation fault error
because of infinite values in the spread. This issue is under
investigation and more information about this segmentation fault will be
documented in a g-w issue. When more information is known about this
issue, the necessary fixes will be made in develop.